Refactor admin models
